refactor: move admin pages into dedicated folder
