+++ client/src/App.jsx
@@ -12,17 +12,17 @@ import Forbidden from './pages/Forbidden';
 import ServerError from './pages/ServerError';
 import MarkdownPage from './pages/MarkdownPage';
 import WidgetPage from './pages/WidgetPage';
-import AdminHome from './pages/AdminHome';
-import AdminUsageReports from './pages/AdminUsageReports';
-import AdminSystemPage from './pages/AdminSystemPage';
-import AdminAppsPage from './pages/AdminAppsPage';
-import AdminAppEditPage from './pages/AdminAppEditPage';
-import AdminShortLinks from './pages/AdminShortLinks';
-import AdminShortLinkEditPage from './pages/AdminShortLinkEditPage';
-import AdminModelEditPage from './pages/AdminModelEditPage';
-import AdminModelsPage from './pages/AdminModelsPage';
-import AdminPromptsPage from './pages/AdminPromptsPage';
-import AdminPromptEditPage from './pages/AdminPromptEditPage';
+import AdminHome from './pages/admin/AdminHome';
+import AdminUsageReports from './pages/admin/AdminUsageReports';
+import AdminSystemPage from './pages/admin/AdminSystemPage';
+import AdminAppsPage from './pages/admin/AdminAppsPage';
+import AdminAppEditPage from './pages/admin/AdminAppEditPage';
+import AdminShortLinks from './pages/admin/AdminShortLinks';
+import AdminShortLinkEditPage from './pages/admin/AdminShortLinkEditPage';
+import AdminModelEditPage from './pages/admin/AdminModelEditPage';
+import AdminModelsPage from './pages/admin/AdminModelsPage';
+import AdminPromptsPage from './pages/admin/AdminPromptsPage';
+import AdminPromptEditPage from './pages/admin/AdminPromptEditPage';
 import AppProviders from './components/AppProviders';
 import { withErrorBoundary } from './components/ErrorBoundary';
 import useSessionManagement from './hooks/useSessionManagement';
