+++ client/src/pages/admin/AdminPromptsPage.jsx
@@ -1,12 +1,12 @@
 import React, { useState, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useTranslation } from 'react-i18next';
-import { getLocalizedContent } from '../utils/localizeContent';
-import Icon from '../components/Icon';
-import AdminAuth from '../components/AdminAuth';
-import AdminNavigation from '../components/AdminNavigation';
-import PromptDetailsPopup from '../components/PromptDetailsPopup';
-import { fetchAdminPrompts, makeAdminApiCall } from '../api/adminApi';
+import { getLocalizedContent } from '../../utils/localizeContent';
+import Icon from '../../components/Icon';
+import AdminAuth from '../../components/AdminAuth';
+import AdminNavigation from '../../components/AdminNavigation';
+import PromptDetailsPopup from '../../components/PromptDetailsPopup';
+import { fetchAdminPrompts, makeAdminApiCall } from '../../api/adminApi';
 
 const AdminPromptsPage = () => {
   const { t, i18n } = useTranslation();
